organizes style sheet folders into page and components
