+++ components/Support/SupportResource.js
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import styles from "../../styles/Support/SupportResource.module.scss";
+import styles from "../../styles/Support/components/SupportResource.module.scss";
 import { urlFor } from "../../lib/sanity";
 
 export const SupportResource = ({ resource }) => {
