+++ pages/index.js
@@ -1,6 +1,6 @@
 import Head from "next/head";
 import Image from "next/image";
-import styles from "../styles/Home/Home.module.scss";
+import styles from "../styles/Home/page/Home.module.scss";
 
 export default function Home() {
   return (
