+++ styles/Support/components/SupportResourcesContainer.module.scss
