+++ styles/Wins/components/WinsResource.module.scss
