+++ styles/Wins/components/WinsResourcesContainer.module.scss
