+++ styles/Wins/page/Wins.module.scss
